+++ src/main/java/fr/hoc/dap/server/data/repository/DapUserRepository.java
@@ -201,9 +201,13 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package fr.hoc.dap.server.data;
+package fr.hoc.dap.server.data.repository;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
+
+import fr.hoc.dap.server.data.entity.DapUser;
 
 /**
  * Manage access for DaP user in database.
@@ -212,5 +216,12 @@ import org.springframework.data.repository.CrudRepository;
  */
 public interface DapUserRepository extends CrudRepository<DapUser, Long> {
 
-    // DapUser findDapUser(String userKey);
+    /**
+     * Get DapUser of loginName.
+     *
+     * @param loginName DaP login Name
+     * @return DapUser
+     */
+    @Query("select user from DapUser user where user.loginName = :loginName")
+    DapUser findByLoginName(@Param("loginName") String loginName);
 }
