+++ src/main/java/fr/hoc/dap/server/data/repository/package-info.java
@@ -201,191 +201,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package fr.hoc.dap.server.controller;
-
-import java.io.IOException;
-import java.security.GeneralSecurityException;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.MissingServletRequestParameterException;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import com.google.api.client.auth.oauth2.AuthorizationCodeRequestUrl;
-import com.google.api.client.auth.oauth2.AuthorizationCodeResponseUrl;
-import com.google.api.client.auth.oauth2.Credential;
-import com.google.api.client.auth.oauth2.TokenResponse;
-import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
-import com.google.api.client.http.GenericUrl;
-
-import fr.hoc.dap.server.service.GoogleService;
-
 /**
- * Manage google account.
+ * Package repository.
  *
  * @author Michel BARRAT && Thomas TAVERNIER
- *
  */
-@Controller
-public class GoogleAccount extends GoogleService {
-    /** logger. */
-    private static final Logger LOG = LogManager.getLogger("GoogleAccount");
-    /** The first caracter to display for senssible datas (like Tokens). */
-    private static final int SENSIBLE_DATA_FIRST_CHAR = 1;
-    /** The last caracter to display for senssible datas (like Tokens). */
-    private static final int SENSIBLE_DATA_LAST_CHAR = 7;
-
-    /**
-     * Retrieve the User ID in Session.
-     *
-     * @param session the HTTP Session.
-     * @return the current User Id in Session.
-     * @throws ServletException if no User Id in session.
-     */
-    private String getUserid(final HttpSession session) throws ServletException {
-        String userId = null;
-        if (null != session && null != session.getAttribute("userId")) {
-            userId = (String) session.getAttribute("userId");
-        }
-        if (null == userId) {
-            LOG.error("userId in Session is NULL in Callback");
-            throw new ServletException("Error when trying to add Google acocunt : userId is NULL is User Session");
-        }
-        return userId;
-    }
-
-    /**
-     * Extract OAuth2 Google code (from URL) and decode it.
-     *
-     * @param request the HTTP request to extract OAuth2 code.
-     * @return the decoded code.
-     * @throws ServletException if the code cannot be decoded.
-     */
-    private String extracCode(final HttpServletRequest request) throws ServletException {
-        final StringBuffer buf = request.getRequestURL();
-        if (null != request.getQueryString()) {
-            buf.append('?').append(request.getQueryString());
-        }
-        final AuthorizationCodeResponseUrl responseUrl = new AuthorizationCodeResponseUrl(buf.toString());
-        final String decodeCode = responseUrl.getCode();
-
-        if (decodeCode == null) {
-            throw new MissingServletRequestParameterException("code", "String");
-        }
-        if (null != responseUrl.getError()) {
-            LOG.error("Error when trying to add Google acocunt : " + responseUrl.getError());
-            throw new ServletException("Error when trying to add Google acocunt");
-
-        }
-        return decodeCode;
-    }
-
-    /**
-     * Build a current host (and port) absolute URL.
-     *
-     * @param req         the current HTTP request to extract schema, host, port informations.
-     * @param destination the "path" to the resource.
-     * @return an absolute URI.
-     */
-    private String buildRedirectUri(final HttpServletRequest req, final String destination) {
-        final GenericUrl url = new GenericUrl(req.getRequestURL().toString());
-        url.setRawPath(destination);
-        return url.build();
-    }
-
-    /**
-     * Handle the Google response.
-     *
-     * @param request the HTTP Request.
-     * @param code    the (encoded) code use by Google (token, expirationDate,...).
-     * @param session the HTTP Session.
-     * @return the view to display.
-     * @throws ServletException         when Google account could not be connected to DaP.
-     * @throws GeneralSecurityException cannot connect to google sever.
-     * @GeneralSecurityException Google problems.
-     */
-    @RequestMapping("/oAuth2Callback")
-    private String oAuthCallback(@RequestParam final String code, final HttpServletRequest request,
-            final HttpSession session) throws ServletException, GeneralSecurityException {
-        final String decodedCode = extracCode(request);
-        final String redirectUri = buildRedirectUri(request, getMyConf().getoAuth2CallbackUrl());
-        final String userId = getUserid(session);
-        try {
-            final GoogleAuthorizationCodeFlow flow = super.getFlow();
-            final TokenResponse response = flow.newTokenRequest(decodedCode).setRedirectUri(redirectUri).execute();
-
-            final Credential credential = flow.createAndStoreCredential(response, userId);
-            if (null == credential || null == credential.getAccessToken()) {
-                LOG.warn("Trying to store a NULL AccessToken for user : " + userId);
-            }
-
-            if (LOG.isDebugEnabled()) {
-                if (null != credential && null != credential.getAccessToken()) {
-                    LOG.debug("New user credential stored with userId : " + userId + "partial AccessToken : "
-                            + credential.getAccessToken().substring(SENSIBLE_DATA_FIRST_CHAR, SENSIBLE_DATA_LAST_CHAR));
-                }
-            }
-        } catch (IOException e) {
-            LOG.error("Exception while trying to store user Credential", e);
-            throw new ServletException("Error while trying to conenct Google Account");
-        }
-        return "redirect:/account/added";
-    }
-
-    /**
-     * Add a Google account (user will be prompt to connect and accept required access).
-     *
-     * @param userId  the user to store Data.
-     * @param request the HTTP request.
-     * @param session the HTTP session.
-     * @return the view to Display (on Error).
-     * @throws GeneralSecurityException cannot connect to google sever.
-     */
-    @RequestMapping("/account/add/{userId}")
-    private String addAccount(@PathVariable final String userId, final HttpServletRequest request,
-            final HttpSession session) throws GeneralSecurityException {
-        String response = "errorOccurs";
-        GoogleAuthorizationCodeFlow flow;
-        Credential credential = null;
-        try {
-            flow = super.getFlow();
-            credential = flow.loadCredential(userId);
-            if (credential != null && credential.getAccessToken() != null) {
-                response = "AccountAlreadyAdded";
-            } else {
-                // redirect to the authorization flow
-                final AuthorizationCodeRequestUrl authorizationUrl = flow.newAuthorizationUrl();
-                authorizationUrl.setRedirectUri(buildRedirectUri(request, getMyConf().getoAuth2CallbackUrl()));
-                // store userId in session for CallBack Access
-                session.setAttribute("userId", userId);
-                response = "redirect:" + authorizationUrl.build();
-            }
-        } catch (IOException e) {
-            LOG.error("Error while loading credential (or Google Flow)", e);
-        }
-        return response;
-    }
-
-    /**
-     * retrieve if account exist.
-     *
-     * @param userKey which user wanted access.
-     * @return account exist true/false.
-     * @throws IOException              if the credentials.json file cannot be found.
-     * @throws GeneralSecurityException cannot connect to google sever.
-     */
-    @RequestMapping("/account/exist")
-    @ResponseBody
-    public Boolean accountNotExist(@RequestParam("userKey") final String userKey)
-            throws IOException, GeneralSecurityException {
-        return super.doesAccountExist(userKey);
-    }
-}
+package fr.hoc.dap.server.data.repository;
