+++ src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/dap_tat
+spring.datasource.username=dap_server
+spring.datasource.password=MKiWiiCmhztmwA2b
+spring.datasource.driverClassName=com.mysql.jdbc.Driver
+spring.jpa.database=MYSQL
+spring.jpa.generate-ddl=true
+spring.datasource.continue-on-error=true
